requirements.txt
streamlit
numpy
tensorflow
pillow
opencv-python
matplotlib
